parallelize chart pulls in store sync and store add chart
