+++ cmd/hauler/cli/store/add_test.go
@@ -1,24 +1,38 @@
 package store
 
 import (
+	"bytes"
+	"context"
+	"errors"
+	"fmt"
+	"io"
 	"net"
+	"net/http"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
 	"reflect"
+	"regexp"
 	"strings"
+	"sync"
+	"sync/atomic"
 	"testing"
+	"time"
 
+	"github.com/dustin/go-humanize"
 	"github.com/google/go-containerregistry/pkg/name"
 	"github.com/google/go-containerregistry/pkg/registry"
 	"github.com/google/go-containerregistry/pkg/v1/remote"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
+	"github.com/rs/zerolog"
+	"helm.sh/helm/v4/pkg/action"
 	helmchart "helm.sh/helm/v4/pkg/chart/v2"
 	"helm.sh/helm/v4/pkg/chart/v2/util"
 
 	"hauler.dev/go/hauler/v2/internal/flags"
 	v1 "hauler.dev/go/hauler/v2/pkg/apis/hauler.cattle.io/v1"
 	"hauler.dev/go/hauler/v2/pkg/consts"
+	"hauler.dev/go/hauler/v2/pkg/store"
 )
 
 // newLocalhostRegistry creates an in-memory OCI registry server listening on
@@ -42,7 +56,7 @@ func newLocalhostRegistry(t *testing.T) (host string, remoteOpts []remote.Option
 }
 
 // chartTestdataDir is the relative path from cmd/hauler/cli/store/ to the
-// top-level testdata directory, matching the convention in chart_test.go.
+// top-level testdata directory, matching the convention in add_test.go.
 // It must remain relative so that url.ParseRequestURI rejects it (an absolute
 // path would be mistakenly treated as a URL by chart.NewChart's isUrl check).
 const chartTestdataDir = "../../../../testdata"
@@ -724,6 +738,7 @@ func TestAddChartCmd_LocalTgz(t *testing.T) {
 	ro := defaultCliOpts()
 
 	o := newAddChartOpts(chartTestdataDir, "")
+	o.Concurrency = consts.DefaultConcurrency
 	if err := AddChartCmd(ctx, o, s, "rancher-cluster-templates-0.5.2.tgz", rso, ro); err != nil {
 		t.Fatalf("AddChartCmd: %v", err)
 	}
@@ -741,10 +756,45 @@ func TestAddChartCmd_WithFileDep(t *testing.T) {
 	ro := defaultCliOpts()
 
 	o := newAddChartOpts(chartTestdataDir, "")
+	o.Concurrency = consts.DefaultConcurrency
 	if err := AddChartCmd(ctx, o, s, "chart-with-file-dependency-chart-1.0.0.tgz", rso, ro); err != nil {
 		t.Fatalf("AddChartCmd: %v", err)
 	}
 	assertArtifactInStore(t, s, "chart-with-file-dependency-chart")
+}
+
+// TestAddChartCmd_DependencyTreeAtVaryingConcurrency proves AddChartCmd
+// resolves a chart's dependency tree correctly regardless of o.Concurrency.
+// It reuses the file-dependency fixture (a parent with two dependencies
+// resolved BFS-style) at both concurrency=1 -- forcing runChartJobs'
+// single-worker path -- and a higher value, and checks the store ends up
+// identical either way, under -race. It does NOT prove concurrency is
+// "honored": nothing here observes actual fan-out or parallelism, so a
+// hardcoded consts.DefaultConcurrency that silently ignored o.Concurrency
+// would still pass both subtests unchanged.
+// TestTraverseChartLevels_BoundedFanOut is the test that actually asserts
+// on the concurrency bound.
+func TestAddChartCmd_DependencyTreeAtVaryingConcurrency(t *testing.T) {
+	for _, concurrency := range []int{1, 4} {
+		t.Run(fmt.Sprintf("concurrency=%d", concurrency), func(t *testing.T) {
+			ctx := newTestContext(t)
+			s := newTestStore(t)
+			rso := defaultRootOpts(s.Root)
+			ro := defaultCliOpts()
+
+			o := newAddChartOpts(chartTestdataDir, "")
+			o.AddDependencies = true
+			o.Concurrency = concurrency
+
+			if err := AddChartCmd(ctx, o, s, "chart-with-file-dependency-chart-1.0.0.tgz", rso, ro); err != nil {
+				t.Fatalf("AddChartCmd concurrency=%d: %v", concurrency, err)
+			}
+
+			assertArtifactInStore(t, s, "chart-with-file-dependency-chart:1.0.0")
+			assertArtifactInStore(t, s, "child:2.0.0")
+			assertArtifactInStore(t, s, "crds:0.0.1")
+		})
+	}
 }
 
 func TestStoreChart_Rewrite(t *testing.T) {
@@ -794,7 +844,7 @@ func seedChartWithImages(t *testing.T, dir string, images []string) string {
 	return saved
 }
 
-func TestStoreChart_AddImages_ExcludeExtras(t *testing.T) {
+func TestRunChartJobs_AddImages_ExcludeExtras(t *testing.T) {
 	ctx := newTestContext(t)
 	host, rOpts := newLocalhostRegistry(t)
 
@@ -812,13 +862,16 @@ func TestStoreChart_AddImages_ExcludeExtras(t *testing.T) {
 	ro := defaultCliOpts()
 
 	t.Run("excludeExtras=true suppresses sigs/atts/sboms for chart-discovered images", func(t *testing.T) {
-		o := &flags.AddChartOpts{
-			ChartOpts:     newAddChartOpts("", "").ChartOpts,
-			AddImages:     true,
-			ExcludeExtras: true,
+		job := chartJob{
+			cfg: v1.Chart{Name: tgzPath},
+			opts: flags.AddChartOpts{
+				ChartOpts:     newAddChartOpts("", "").ChartOpts,
+				AddImages:     true,
+				ExcludeExtras: true,
+			},
 		}
-		if err := storeChart(ctx, s, v1.Chart{Name: tgzPath}, o, rso, ro, ""); err != nil {
-			t.Fatalf("storeChart with ExcludeExtras: %v", err)
+		if err := runChartJobs(ctx, s, []chartJob{job}, 1, rso, ro, nil); err != nil {
+			t.Fatalf("runChartJobs with ExcludeExtras: %v", err)
 		}
 
 		// The chart itself is stored as an OCI image artifact.
@@ -844,7 +897,7 @@ func TestStoreChart_AddImages_ExcludeExtras(t *testing.T) {
 	})
 }
 
-func TestStoreChart_AddImages_IncludeExtras(t *testing.T) {
+func TestRunChartJobs_AddImages_IncludeExtras(t *testing.T) {
 	ctx := newTestContext(t)
 	host, rOpts := newLocalhostRegistry(t)
 
@@ -861,13 +914,16 @@ func TestStoreChart_AddImages_IncludeExtras(t *testing.T) {
 	ro := defaultCliOpts()
 
 	t.Run("excludeExtras=false includes sigs/atts/sboms for chart-discovered images", func(t *testing.T) {
-		o := &flags.AddChartOpts{
-			ChartOpts:     newAddChartOpts("", "").ChartOpts,
-			AddImages:     true,
-			ExcludeExtras: false,
+		job := chartJob{
+			cfg: v1.Chart{Name: tgzPath},
+			opts: flags.AddChartOpts{
+				ChartOpts:     newAddChartOpts("", "").ChartOpts,
+				AddImages:     true,
+				ExcludeExtras: false,
+			},
 		}
-		if err := storeChart(ctx, s, v1.Chart{Name: tgzPath}, o, rso, ro, ""); err != nil {
-			t.Fatalf("storeChart without ExcludeExtras: %v", err)
+		if err := runChartJobs(ctx, s, []chartJob{job}, 1, rso, ro, nil); err != nil {
+			t.Fatalf("runChartJobs without ExcludeExtras: %v", err)
 		}
 
 		assertArtifactKindInStore(t, s, "test/chart-image:v2", consts.KindAnnotationSigs)
@@ -880,16 +936,16 @@ func TestStoreChart_AddImages_IncludeExtras(t *testing.T) {
 // --local flag validation tests
 // --------------------------------------------------------------------------
 
-// TestStoreChart_AddImages_IgnoreErrors_EnvVar verifies that a chart-discovered
-// image failure inside storeChart's images loop is swallowed via
-// HAULER_IGNORE_ERRORS alone, without --ignore-errors, and that storeChart
-// continues on to store the chart itself. Regression test: retry.Operation and
-// storeImage used to mutate the shared ro.IgnoreErrors from the env var as a
-// side effect, so ro.IgnoreErrors reads later in the images loop always saw the
-// env var. Now that storeImage/retry.Operation are pure reads via
-// flags.ShouldIgnoreErrors, the remaining ro.IgnoreErrors reads in the loop must
-// still honor the env var on their own.
-func TestStoreChart_AddImages_IgnoreErrors_EnvVar(t *testing.T) {
+// TestRunChartJobs_AddImages_IgnoreErrors_EnvVar verifies that a
+// chart-discovered image failure is swallowed via HAULER_IGNORE_ERRORS alone,
+// without --ignore-errors, and that the chart itself is still stored.
+// Regression test: retry.Operation and storeImage used to mutate the shared
+// ro.IgnoreErrors from the env var as a side effect, so every ro.IgnoreErrors
+// read after the first one saw the env var. Now that storeImage and
+// retry.Operation are pure reads via flags.ShouldIgnoreErrors, fetchChart's
+// image-discovery loop and the image phase must each honor the env var on
+// their own.
+func TestRunChartJobs_AddImages_IgnoreErrors_EnvVar(t *testing.T) {
 	ctx := newTestContext(t)
 
 	t.Run("applyDefaultRegistry failure via env var", func(t *testing.T) {
@@ -904,16 +960,19 @@ func TestStoreChart_AddImages_IgnoreErrors_EnvVar(t *testing.T) {
 		ro := defaultCliOpts()
 		t.Setenv(consts.HaulerIgnoreErrors, "true")
 
-		o := &flags.AddChartOpts{
-			ChartOpts: newAddChartOpts("", "").ChartOpts,
-			AddImages: true,
-			Registry:  "registry.example.com",
+		job := chartJob{
+			cfg: v1.Chart{Name: tgzPath},
+			opts: flags.AddChartOpts{
+				ChartOpts: newAddChartOpts("", "").ChartOpts,
+				AddImages: true,
+				Registry:  "registry.example.com",
+			},
 		}
-		if err := storeChart(ctx, s, v1.Chart{Name: tgzPath}, o, rso, ro, ""); err != nil {
+		if err := runChartJobs(ctx, s, []chartJob{job}, 1, rso, ro, nil); err != nil {
 			t.Fatalf("expected nil with HAULER_IGNORE_ERRORS=true, got: %v", err)
 		}
 		if ro.IgnoreErrors {
-			t.Fatal("expected ro.IgnoreErrors to remain false: storeChart must not mutate ro")
+			t.Fatal("expected ro.IgnoreErrors to remain false: the chart path must not mutate ro")
 		}
 		assertArtifactInStore(t, s, "test-chart")
 	})
@@ -928,15 +987,18 @@ func TestStoreChart_AddImages_IgnoreErrors_EnvVar(t *testing.T) {
 		ro := defaultCliOpts()
 		t.Setenv(consts.HaulerIgnoreErrors, "true")
 
-		o := &flags.AddChartOpts{
-			ChartOpts: newAddChartOpts("", "").ChartOpts,
-			AddImages: true,
+		job := chartJob{
+			cfg: v1.Chart{Name: tgzPath},
+			opts: flags.AddChartOpts{
+				ChartOpts: newAddChartOpts("", "").ChartOpts,
+				AddImages: true,
+			},
 		}
-		if err := storeChart(ctx, s, v1.Chart{Name: tgzPath}, o, rso, ro, ""); err != nil {
+		if err := runChartJobs(ctx, s, []chartJob{job}, 1, rso, ro, nil); err != nil {
 			t.Fatalf("expected nil with HAULER_IGNORE_ERRORS=true, got: %v", err)
 		}
 		if ro.IgnoreErrors {
-			t.Fatal("expected ro.IgnoreErrors to remain false: storeChart must not mutate ro")
+			t.Fatal("expected ro.IgnoreErrors to remain false: the chart path must not mutate ro")
 		}
 		assertArtifactInStore(t, s, "test-chart")
 	})
@@ -1034,4 +1096,1634 @@ func TestStoreLocalImage_InvalidReference(t *testing.T) {
 			t.Fatal("expected ro.IgnoreErrors to remain false: storeLocalImage must not mutate ro")
 		}
 	})
+}
+
+// --------------------------------------------------------------------------
+// Durable index save tests
+// --------------------------------------------------------------------------
+
+// add_durable_index_test.go covers the durability gap identified in the
+// final cross-task review of the I/O tuning workstream: index.json's
+// per-artifact fsync was replaced with a 30-second durable checkpoint
+// (content.OCI.saveIndexCheckpointLocked), so nothing below the command
+// entry points forces an fsync any more. Ending a run durably is now the
+// job of each entry point's deferred SaveIndex() -- AddFileCmd, AddImageCmd,
+// and AddChartCmd each have one. Without it a completed `hauler store add
+// image`/`add file`/`add chart` could return success with its final
+// index.json state sitting only in page cache, not fsynced -- a
+// lost-index-entry risk on crash/power-loss (blobs are unaffected;
+// writeBlobOnce always fsyncs).
+//
+// These tests observe durability via content.OCI.Stats().Snapshot()'s
+// IndexDurableWrites counter rather than by inspecting the filesystem
+// directly, since "was this fsync'd" isn't otherwise observable from
+// outside the package.
+
+// TestAddImageCmd_EndsWithDurableIndexSave reproduces the gap for
+// AddImageCmd. store.Layout.AddImage calls content.OCI.AddIndex once for
+// the base image, then once more per discovered cosign signature,
+// attestation, and SBOM (see saveRelatedArtifacts) -- all against the same
+// store instance, well within the 30s checkpoint window. Because
+// content.OCI's lastDurableSave starts at its zero value, Since(zero) is
+// always >= 30s, so the very first AddIndex call of a fresh store is
+// durable "for free" -- but every subsequent call, including the last one
+// that actually reflects the complete set of discovered artifacts, is not.
+// A fix must add a trailing durable SaveIndex() so the run always ends
+// durable regardless of how many AddIndex calls happened inside it.
+func TestAddImageCmd_EndsWithDurableIndexSave(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+
+	host, remoteOpts := newLocalhostRegistry(t)
+	img := seedImage(t, host, "myorg/durable", "v1", remoteOpts...)
+	seedCosignV2Artifacts(t, host, "myorg/durable", img, remoteOpts...)
+
+	o := &flags.AddImageOpts{}
+	rso := defaultRootOpts(s.Root)
+	ro := defaultCliOpts()
+
+	ref := fmt.Sprintf("%s/myorg/durable:v1", host)
+	if err := AddImageCmd(ctx, o, s, ref, rso, ro); err != nil {
+		t.Fatalf("AddImageCmd: %v", err)
+	}
+
+	snap := s.OCI.Stats().Snapshot()
+	// Without the fix this is 1: only the very first AddIndex call (image
+	// itself) lands durably, courtesy of lastDurableSave's zero value. The
+	// sig/att/sbom AddIndex calls that follow -- and the run's final state
+	// -- are not durable until a trailing SaveIndex() is added.
+	if snap.IndexDurableWrites < 2 {
+		t.Fatalf("expected at least 2 durable index writes (initial AddIndex + trailing checkpoint), got %d (total index writes=%d)", snap.IndexDurableWrites, snap.IndexWrites)
+	}
+}
+
+// TestAddFileCmd_EndsWithDurableIndexSave verifies AddFileCmd also ends its
+// run with a durable index save. A single AddFileCmd call only performs one
+// AddIndex call, which (per TestAddImageCmd_EndsWithDurableIndexSave's
+// rationale) is already durable "for free" on a brand new store -- so this
+// test seeds an unrelated file first to consume that free durability and
+// set lastDurableSave to "now", then immediately adds a second file so its
+// AddIndex call falls inside the 30s checkpoint window and would be
+// non-durable without the fix.
+func TestAddFileCmd_EndsWithDurableIndexSave(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+
+	warmup, err := os.CreateTemp(t.TempDir(), "warmup-*.txt")
+	if err != nil {
+		t.Fatal(err)
+	}
+	warmup.WriteString("warmup content") //nolint:errcheck
+	warmup.Close()
+
+	warmupOpts := &flags.AddFileOpts{StoreRootOpts: defaultRootOpts(s.Root)}
+	if err := AddFileCmd(ctx, warmupOpts, s, warmup.Name(), defaultCliOpts()); err != nil {
+		t.Fatalf("AddFileCmd warmup: %v", err)
+	}
+
+	tmp, err := os.CreateTemp(t.TempDir(), "durable-*.txt")
+	if err != nil {
+		t.Fatal(err)
+	}
+	tmp.WriteString("durable content") //nolint:errcheck
+	tmp.Close()
+
+	before := s.OCI.Stats().Snapshot().IndexDurableWrites
+
+	o := &flags.AddFileOpts{StoreRootOpts: defaultRootOpts(s.Root)}
+	if err := AddFileCmd(ctx, o, s, tmp.Name(), defaultCliOpts()); err != nil {
+		t.Fatalf("AddFileCmd: %v", err)
+	}
+
+	after := s.OCI.Stats().Snapshot().IndexDurableWrites
+	if after <= before {
+		t.Fatalf("expected a trailing durable index save after AddFileCmd, durable writes went from %d to %d", before, after)
+	}
+}
+
+// --------------------------------------------------------------------------
+// File retry tests
+// --------------------------------------------------------------------------
+
+// add_file_retry_test.go covers storeFile's (cmd/hauler/cli/store/add.go)
+// retry, --ignore-errors, and cancellation behavior -- extending it to match
+// storeImage's existing shape (see add_retry_stats_test.go for the analogous
+// image-side retry test) as part of bringing Files up to parity with Images
+// for `hauler store sync`'s --concurrency support.
+
+// TestStoreFile_RetriesOnTransientFailure proves storeFile retries a failed
+// fetch via retry.Operation rather than aborting the whole sync on one
+// transient HTTP blip -- storeFile previously had no retry wrapping at all.
+func TestStoreFile_RetriesOnTransientFailure(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping: requires one RetriesInterval sleep (5s)")
+	}
+
+	var gets int32
+	mux := http.NewServeMux()
+	mux.HandleFunc("/flaky.sh", func(w http.ResponseWriter, r *http.Request) {
+		// storeFile's Client.Name(fi.Path) call (used to derive the stored
+		// ref, before retry.Operation ever starts) issues an unconditional
+		// HEAD request -- see getter.Http.Name -- that must not count
+		// against the GET-failure budget below, or the "failure" would be
+		// silently consumed before AddArtifact's first real attempt.
+		if r.Method != http.MethodGet {
+			w.WriteHeader(http.StatusOK)
+			return
+		}
+		if atomic.AddInt32(&gets, 1) == 1 {
+			w.WriteHeader(http.StatusInternalServerError)
+			return
+		}
+		io.WriteString(w, "#!/bin/sh\necho ok") //nolint:errcheck
+	})
+	srv := httptest.NewServer(mux)
+	t.Cleanup(srv.Close)
+
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+
+	rso := defaultRootOpts(s.Root)
+	rso.Retries = 2 // one failed attempt + one successful retry
+	ro := defaultCliOpts()
+
+	if err := storeFile(ctx, s, v1.File{Path: srv.URL + "/flaky.sh"}, ro, rso); err != nil {
+		t.Fatalf("storeFile: %v", err)
+	}
+	assertArtifactInStore(t, s, "flaky.sh")
+	// The first GET fails; layer.FromOpener opens once per successful
+	// LayerFrom call (it derives diffID from the already-computed digest
+	// instead of a second read), so a successful retry attempt adds 1 more
+	// -- 2 total.
+	if got := atomic.LoadInt32(&gets); got < 2 {
+		t.Errorf("expected at least 2 GET attempts (1 failure + 1 for the succeeding retry), got %d", got)
+	}
+}
+
+// TestStoreFile_IgnoreErrors_WarnsAndReturnsNil proves storeFile absorbs a
+// failure into a warning and returns nil when --ignore-errors is set,
+// matching storeImage's existing behavior -- previously storeFile always
+// propagated the error regardless of ignore-errors.
+func TestStoreFile_IgnoreErrors_WarnsAndReturnsNil(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+
+	ro := defaultCliOpts()
+	ro.IgnoreErrors = true
+	rso := defaultRootOpts(s.Root)
+
+	err := storeFile(ctx, s, v1.File{Path: "/nonexistent/path/missing-file.txt"}, ro, rso)
+	if err != nil {
+		t.Fatalf("expected nil error with --ignore-errors, got: %v", err)
+	}
+	if n := countArtifactsInStore(t, s); n != 0 {
+		t.Errorf("expected 0 artifacts after an ignored failure, got %d", n)
+	}
+}
+
+// TestStoreFile_ContextAlreadyCancelled_ReturnsPromptly is the regression
+// test for File.compute()'s context.TODO() bug (pkg/artifacts/file/file.go):
+// storeFile must check ctx and bail out before ever attempting to fetch,
+// matching storeImage's early ctx.Err() check.
+func TestStoreFile_ContextAlreadyCancelled_ReturnsPromptly(t *testing.T) {
+	s := newTestStore(t)
+
+	zl := zerolog.New(io.Discard)
+	ctx, cancel := context.WithCancel(zl.WithContext(context.Background()))
+	cancel()
+
+	ro := defaultCliOpts()
+	rso := defaultRootOpts(s.Root)
+
+	err := storeFile(ctx, s, v1.File{Path: "https://example.invalid/never-fetched.sh"}, ro, rso)
+	if err == nil {
+		t.Fatal("expected an error for an already-cancelled context, got nil")
+	}
+	if n := countArtifactsInStore(t, s); n != 0 {
+		t.Errorf("expected 0 artifacts, got %d", n)
+	}
+}
+
+// TestStoreFile_CompletionLine_Format proves storeFile logs a "✓ added"
+// completion line (matching storeImage's formatAddedLine convention) rather
+// than the old plain "successfully added file" line, and that the old
+// "adding file" line is demoted to debug (absent at the default/error log
+// level defaultCliOpts() uses).
+func TestStoreFile_CompletionLine_Format(t *testing.T) {
+	url := seedFileInHTTPServer(t, "completion.sh", "#!/bin/sh\necho done")
+
+	s := newTestStore(t)
+	var buf bytes.Buffer
+	zl := zerolog.New(&buf).Level(zerolog.InfoLevel)
+	ctx := zl.WithContext(context.Background())
+
+	ro := defaultCliOpts()
+	ro.LogLevel = "info"
+	rso := defaultRootOpts(s.Root)
+
+	if err := storeFile(ctx, s, v1.File{Path: url}, ro, rso); err != nil {
+		t.Fatalf("storeFile: %v", err)
+	}
+
+	out := buf.String()
+	if !strings.Contains(out, "✓ added") {
+		t.Errorf("expected a \"✓ added\" completion line, got:\n%s", out)
+	}
+	if strings.Contains(out, "successfully added file") {
+		t.Errorf("expected the old \"successfully added file\" line to be gone, got:\n%s", out)
+	}
+}
+
+// --------------------------------------------------------------------------
+// Image retry stats tests
+// --------------------------------------------------------------------------
+
+// add_retry_stats_test.go covers a single narrow bug in storeImage
+// (cmd/hauler/cli/store/add.go): a retried s.AddImage attempt must not
+// accumulate layer count / byte totals onto the same store.ImageStats
+// pointer used by a prior, failed attempt -- otherwise the "✓ added ..."
+// completion line double (or N-x) counts layers/bytes across retries.
+
+// failOnceForDigestHandler wraps an http.Handler and fails the very first GET
+// request for one specific blob digest with a 403, then lets every other
+// request -- including every later request for that same digest -- through
+// unmodified. Targeting a single, specific digest (rather than "the first
+// request for whichever digest shows up first") is deliberate: writeImageBlobs
+// fetches an image's layers concurrently and its config sequentially
+// afterward, so a naive "fail every digest's first-ever request" approach
+// makes the *retry* attempt also fail (on whichever blob it reaches first
+// that the earlier, aborted attempt never got around to requesting at all).
+// Failing exactly one predetermined digest, exactly once, guarantees the
+// first AddImage attempt fails (on that blob) while the second, retried
+// attempt succeeds outright (every blob, including the previously-failing
+// one, now passes through).
+//
+// 403 (rather than 500/502/503/504/408/429) is deliberate: go-containerregistry's
+// remote transport has its own built-in retry for a fixed set of "temporary"
+// status codes (see remote.defaultRetryStatusCodes) and would silently absorb
+// a transient 500 within a single AddImage call (up to 3 internal attempts).
+// 403 isn't in that set, so it surfaces as a hard error from that AddImage
+// call and exercises this package's own retry.Operation-driven retry instead.
+type failOnceForDigestHandler struct {
+	next   http.Handler
+	target string // e.g. "sha256:abcd..."
+
+	mu     sync.Mutex
+	failed bool
+}
+
+func (f *failOnceForDigestHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	if r.Method == http.MethodGet && strings.HasSuffix(r.URL.Path, "/blobs/"+f.target) {
+		f.mu.Lock()
+		alreadyFailed := f.failed
+		f.failed = true
+		f.mu.Unlock()
+
+		if !alreadyFailed {
+			w.WriteHeader(http.StatusForbidden)
+			return
+		}
+	}
+	f.next.ServeHTTP(w, r)
+}
+
+// newFailOnceRegistry starts an in-memory OCI registry, listening on
+// "localhost:0" (so go-containerregistry auto-selects plain HTTP, matching
+// newLocalhostRegistry in add_test.go), wrapped in a failOnceForDigestHandler.
+// The returned handler's target field must be set (to the digest that should
+// fail exactly once) before the read that should fail; that happens after
+// seeding, once the seeded image's layer digest is known, and before the
+// caller reads the image back via AddImage/storeImage.
+func newFailOnceRegistry(t *testing.T) (host string, remoteOpts []remote.Option, handler *failOnceForDigestHandler) {
+	t.Helper()
+	handler = &failOnceForDigestHandler{next: registry.New()}
+
+	l, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		t.Fatalf("newFailOnceRegistry listen: %v", err)
+	}
+	srv := httptest.NewUnstartedServer(handler)
+	srv.Listener = l
+	srv.Start()
+	t.Cleanup(srv.Close)
+	host = strings.TrimPrefix(srv.URL, "http://")
+	remoteOpts = []remote.Option{remote.WithTransport(srv.Client().Transport)}
+	return host, remoteOpts, handler
+}
+
+// addedLineLayerCount extracts the layer count from the last "✓ added ...
+// (N layer(s), ...)" line in out. Fails the test if no such line is found.
+func addedLineLayerCount(t *testing.T, out string) int {
+	t.Helper()
+	re := regexp.MustCompile(`✓ added .*\((\d+) layers?,`)
+	matches := re.FindAllStringSubmatch(out, -1)
+	if len(matches) == 0 {
+		t.Fatalf("no \"✓ added ... (N layer(s), ...)\" line found in output:\n%s", out)
+	}
+	last := matches[len(matches)-1]
+	n := 0
+	for _, c := range last[1] {
+		n = n*10 + int(c-'0')
+	}
+	return n
+}
+
+// TestStoreImage_RetryDoesNotDoubleCountStats is the regression test for the
+// ImageStats double-counting bug: storeImage built a single store.ImageStats
+// pointer outside retry.Operation's closure, so a failed first attempt that
+// had already accumulated layer/byte counts into it left those counts in
+// place for a subsequent, successful retry attempt to accumulate on top of --
+// inflating the final completion line's layer/byte totals.
+//
+// seedImage (testhelpers_test.go) always creates a 2-layer image. A single,
+// uncontested successful AddImage call reports "2 layers"; the bug would
+// report "4 layers" after exactly one failed attempt followed by one
+// successful retry.
+func TestStoreImage_RetryDoesNotDoubleCountStats(t *testing.T) {
+	if testing.Short() {
+		t.Skip("skipping: requires one RetriesInterval sleep (5s)")
+	}
+
+	host, remoteOpts, handler := newFailOnceRegistry(t)
+	seededImg := seedImage(t, host, "test/retry-stats", "v1", remoteOpts...)
+
+	layers, err := seededImg.Layers()
+	if err != nil {
+		t.Fatalf("seededImg.Layers: %v", err)
+	}
+	if len(layers) == 0 {
+		t.Fatal("seeded image has no layers")
+	}
+	targetDigest, err := layers[0].Digest()
+	if err != nil {
+		t.Fatalf("layers[0].Digest: %v", err)
+	}
+	handler.target = targetDigest.String()
+
+	s := newTestStore(t)
+	var buf bytes.Buffer
+	zl := zerolog.New(&buf)
+	ctx := zl.WithContext(context.Background())
+
+	rso := defaultRootOpts(s.Root)
+	rso.Retries = 2 // one failed attempt + one successful retry
+	ro := defaultCliOpts()
+
+	cfg := v1.Image{Name: host + "/test/retry-stats:v1"}
+	if err := storeImage(ctx, s, cfg, "", true /* excludeExtras: keep this to just the image's own layers */, rso, ro, ""); err != nil {
+		t.Fatalf("storeImage: %v", err)
+	}
+
+	out := buf.String()
+	if got := addedLineLayerCount(t, out); got != 2 {
+		t.Errorf("reported layer count = %d, want 2 (seedImage's fixed layer count); a retried attempt must not double-count onto the prior failed attempt's stats\nfull output:\n%s", got, out)
+	}
+}
+
+// --------------------------------------------------------------------------
+// resolveChartJobs tests
+//
+// These exercise the Charts precedence rules directly, without a store or
+// network access. Credential resolution (UsernameEnv/PasswordEnv and the
+// field passthrough onto ChartOpts) is covered separately by
+// TestResolveChartJobs_CredentialFields, TestResolveChartJobs_CredentialEnv,
+// TestResolveChartJobs_CredentialEnvMismatch, and
+// TestResolveChartJobs_CredentialIsolation below.
+// --------------------------------------------------------------------------
+
+func TestResolveChartJobs_Registry(t *testing.T) {
+	tests := []struct {
+		name        string
+		cliRegistry string
+		annotation  string
+		want        string
+	}{
+		{
+			name:        "CLI flag wins over annotation",
+			cliRegistry: "cli-registry.io",
+			annotation:  "annotation-registry.io",
+			want:        "cli-registry.io",
+		},
+		{
+			name:       "annotation used when no CLI flag",
+			annotation: "annotation-registry.io",
+			want:       "annotation-registry.io",
+		},
+		{
+			name:        "CLI flag used when no annotation",
+			cliRegistry: "cli-registry.io",
+			want:        "cli-registry.io",
+		},
+		{
+			name: "neither set leaves registry empty",
+			want: "",
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			o := &flags.SyncOpts{Registry: tc.cliRegistry}
+			a := map[string]string{}
+			if tc.annotation != "" {
+				a[consts.ImageAnnotationRegistry] = tc.annotation
+			}
+
+			jobs, err := resolveChartJobs(o, a, "/manifests", []v1.Chart{{Name: "rancher"}})
+			if err != nil {
+				t.Fatalf("resolveChartJobs: %v", err)
+			}
+			if len(jobs) != 1 {
+				t.Fatalf("expected 1 job, got %d", len(jobs))
+			}
+			if got := jobs[0].opts.Registry; got != tc.want {
+				t.Errorf("registry = %q, want %q", got, tc.want)
+			}
+		})
+	}
+}
+
+func TestResolveChartJobs_ExcludeExtras(t *testing.T) {
+	tests := []struct {
+		name       string
+		cli        bool
+		annotation string
+		perChart   bool
+		want       bool
+	}{
+		{name: "nothing set", want: false},
+		{name: "CLI flag alone", cli: true, want: true},
+		{name: "annotation alone", annotation: "true", want: true},
+		{name: "per-chart alone", perChart: true, want: true},
+		{
+			name:       "annotation only honored when set to the literal true",
+			annotation: "yes",
+			want:       false,
+		},
+		{
+			// Neither the annotation nor the per-chart field can turn a CLI
+			// --exclude-extras back off; both are one-way switches.
+			name:       "CLI flag survives an annotation that is not true",
+			cli:        true,
+			annotation: "false",
+			want:       true,
+		},
+		{
+			name:     "CLI flag survives a false per-chart field",
+			cli:      true,
+			perChart: false,
+			want:     true,
+		},
+		{
+			name:       "annotation survives a false per-chart field",
+			annotation: "true",
+			perChart:   false,
+			want:       true,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			o := &flags.SyncOpts{ExcludeExtras: tc.cli}
+			a := map[string]string{}
+			if tc.annotation != "" {
+				a[consts.ImageAnnotationExcludeExtras] = tc.annotation
+			}
+
+			jobs, err := resolveChartJobs(o, a, "/manifests", []v1.Chart{{Name: "rancher", ExcludeExtras: tc.perChart}})
+			if err != nil {
+				t.Fatalf("resolveChartJobs: %v", err)
+			}
+			if len(jobs) != 1 {
+				t.Fatalf("expected 1 job, got %d", len(jobs))
+			}
+			if got := jobs[0].opts.ExcludeExtras; got != tc.want {
+				t.Errorf("excludeExtras = %v, want %v", got, tc.want)
+			}
+		})
+	}
+}
+
+func TestResolveChartJobs_Platform(t *testing.T) {
+	tests := []struct {
+		name       string
+		cli        string
+		annotation string
+		perChart   string
+		want       string
+	}{
+		{name: "nothing set", want: ""},
+		{name: "CLI flag alone", cli: "linux/amd64", want: "linux/amd64"},
+		{name: "annotation alone", annotation: "linux/arm64", want: "linux/arm64"},
+		{name: "per-chart alone", perChart: "linux/s390x", want: "linux/s390x"},
+		{
+			// An explicit --platform is run-time intent and outranks manifest
+			// metadata, matching resolveImageJobs and matching registry's
+			// CLI-over-annotation rule in this same resolver.
+			name:       "CLI flag wins over annotation",
+			cli:        "linux/amd64",
+			annotation: "linux/arm64",
+			want:       "linux/amd64",
+		},
+		{
+			name:       "per-chart wins over both",
+			cli:        "linux/amd64",
+			annotation: "linux/arm64",
+			perChart:   "linux/s390x",
+			want:       "linux/s390x",
+		},
+		{
+			name:       "empty annotation leaves the CLI flag intact",
+			cli:        "linux/amd64",
+			annotation: "",
+			want:       "linux/amd64",
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			o := &flags.SyncOpts{Platform: tc.cli}
+			a := map[string]string{}
+			if tc.annotation != "" {
+				a[consts.ImageAnnotationPlatform] = tc.annotation
+			}
+
+			jobs, err := resolveChartJobs(o, a, "/manifests", []v1.Chart{{Name: "rancher", Platform: tc.perChart}})
+			if err != nil {
+				t.Fatalf("resolveChartJobs: %v", err)
+			}
+			if len(jobs) != 1 {
+				t.Fatalf("expected 1 job, got %d", len(jobs))
+			}
+			if got := jobs[0].opts.Platform; got != tc.want {
+				t.Errorf("platform = %v, want %v", got, tc.want)
+			}
+		})
+	}
+}
+
+// TestResolveChartJobs_PlatformMatchesImagePath pins the two resolvers
+// together: a manifest carrying a platform annotation must select the same
+// platform for its Charts and its Images given the same flags. The two
+// resolvers hold separate copies of this precedence chain, and drift between
+// them is exactly the bug this pins against.
+func TestResolveChartJobs_PlatformMatchesImagePath(t *testing.T) {
+	tests := []struct {
+		name       string
+		cli        string
+		annotation string
+		perEntry   string
+	}{
+		{name: "nothing set"},
+		{name: "CLI flag alone", cli: "linux/amd64"},
+		{name: "annotation alone", annotation: "linux/arm64"},
+		{name: "per-entry alone", perEntry: "linux/s390x"},
+		{name: "CLI flag and annotation", cli: "linux/amd64", annotation: "linux/arm64"},
+		{name: "per-entry and annotation", annotation: "linux/arm64", perEntry: "linux/s390x"},
+		{name: "per-entry and CLI flag", cli: "linux/amd64", perEntry: "linux/s390x"},
+		{name: "all three", cli: "linux/amd64", annotation: "linux/arm64", perEntry: "linux/s390x"},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			o := &flags.SyncOpts{Platform: tc.cli}
+			a := map[string]string{}
+			if tc.annotation != "" {
+				a[consts.ImageAnnotationPlatform] = tc.annotation
+			}
+
+			chartJobs, err := resolveChartJobs(o, a, "/manifests", []v1.Chart{{Name: "rancher", Platform: tc.perEntry}})
+			if err != nil {
+				t.Fatalf("resolveChartJobs: %v", err)
+			}
+			if len(chartJobs) != 1 {
+				t.Fatalf("expected 1 chart job, got %d", len(chartJobs))
+			}
+
+			imageJobs, err := resolveImageJobs(o, a, []v1.Image{{Name: "rancher/rancher:v2.9", Platform: tc.perEntry}})
+			if err != nil {
+				t.Fatalf("resolveImageJobs: %v", err)
+			}
+			if len(imageJobs) != 1 {
+				t.Fatalf("expected 1 image job, got %d", len(imageJobs))
+			}
+
+			if chartJobs[0].opts.Platform != imageJobs[0].platform {
+				t.Errorf("chart platform = %q, image platform = %q, want them equal",
+					chartJobs[0].opts.Platform, imageJobs[0].platform)
+			}
+		})
+	}
+}
+
+func TestResolveChartJobs_NilAnnotations(t *testing.T) {
+	o := &flags.SyncOpts{Registry: "cli-registry.io", Platform: "linux/amd64", ExcludeExtras: true}
+
+	jobs, err := resolveChartJobs(o, nil, "/manifests", []v1.Chart{{Name: "rancher"}})
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 1 {
+		t.Fatalf("expected 1 job, got %d", len(jobs))
+	}
+	if got := jobs[0].opts.Registry; got != "cli-registry.io" {
+		t.Errorf("registry = %q, want cli-registry.io", got)
+	}
+	if got := jobs[0].opts.Platform; got != "linux/amd64" {
+		t.Errorf("platform = %q, want linux/amd64", got)
+	}
+	if !jobs[0].opts.ExcludeExtras {
+		t.Error("excludeExtras = false, want true")
+	}
+}
+
+func TestResolveChartJobs_ValuesFiles(t *testing.T) {
+	tests := []struct {
+		name        string
+		manifestDir string
+		valuesFiles []string
+		want        []string
+	}{
+		{
+			name:        "relative paths join against the manifest directory",
+			manifestDir: "/manifests",
+			valuesFiles: []string{"values.yaml", "overrides/extra.yaml"},
+			want:        []string{"/manifests/values.yaml", "/manifests/overrides/extra.yaml"},
+		},
+		{
+			name:        "parent-relative paths are cleaned",
+			manifestDir: "/manifests/prod",
+			valuesFiles: []string{"../shared/values.yaml"},
+			want:        []string{"/manifests/shared/values.yaml"},
+		},
+		{
+			name:        "manifest in the current directory",
+			manifestDir: ".",
+			valuesFiles: []string{"values.yaml"},
+			want:        []string{"values.yaml"},
+		},
+		{
+			name:        "no values files yields none",
+			manifestDir: "/manifests",
+			valuesFiles: nil,
+			want:        nil,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, tc.manifestDir,
+				[]v1.Chart{{Name: "rancher", ValuesFiles: tc.valuesFiles}})
+			if err != nil {
+				t.Fatalf("resolveChartJobs: %v", err)
+			}
+			if len(jobs) != 1 {
+				t.Fatalf("expected 1 job, got %d", len(jobs))
+			}
+
+			got := jobs[0].opts.ValuesFiles
+			if len(got) != len(tc.want) {
+				t.Fatalf("valuesFiles = %v, want %v", got, tc.want)
+			}
+			for i := range tc.want {
+				if got[i] != tc.want[i] {
+					t.Errorf("valuesFiles[%d] = %q, want %q", i, got[i], tc.want[i])
+				}
+			}
+		})
+	}
+}
+
+func TestResolveChartJobs_PerChartFields(t *testing.T) {
+	charts := []v1.Chart{
+		{
+			Name:            "rancher",
+			RepoURL:         "https://releases.rancher.com/server-charts/stable",
+			Version:         "2.9.0",
+			Rewrite:         "mirror/rancher",
+			AddImages:       true,
+			AddDependencies: true,
+		},
+		{
+			Name:    "cert-manager",
+			RepoURL: "https://charts.jetstack.io",
+			Version: "1.15.0",
+		},
+	}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", charts)
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 2 {
+		t.Fatalf("expected 2 jobs, got %d", len(jobs))
+	}
+
+	if jobs[0].cfg.Name != "rancher" || jobs[1].cfg.Name != "cert-manager" {
+		t.Fatalf("jobs are out of manifest order: %q, %q", jobs[0].cfg.Name, jobs[1].cfg.Name)
+	}
+
+	if jobs[0].rewrite != "mirror/rancher" {
+		t.Errorf("jobs[0].rewrite = %q, want mirror/rancher", jobs[0].rewrite)
+	}
+	if jobs[1].rewrite != "" {
+		t.Errorf("jobs[1].rewrite = %q, want empty", jobs[1].rewrite)
+	}
+	if !jobs[0].opts.AddImages || !jobs[0].opts.AddDependencies {
+		t.Errorf("jobs[0] add-images/add-dependencies = %v/%v, want true/true", jobs[0].opts.AddImages, jobs[0].opts.AddDependencies)
+	}
+	if jobs[1].opts.AddImages || jobs[1].opts.AddDependencies {
+		t.Errorf("jobs[1] add-images/add-dependencies = %v/%v, want false/false", jobs[1].opts.AddImages, jobs[1].opts.AddDependencies)
+	}
+
+	for i, want := range charts {
+		if got := jobs[i].opts.ChartOpts.RepoURL; got != want.RepoURL {
+			t.Errorf("jobs[%d].opts.ChartOpts.RepoURL = %q, want %q", i, got, want.RepoURL)
+		}
+		if got := jobs[i].opts.ChartOpts.Version; got != want.Version {
+			t.Errorf("jobs[%d].opts.ChartOpts.Version = %q, want %q", i, got, want.Version)
+		}
+		if jobs[i].parent != "" {
+			t.Errorf("jobs[%d].parent = %q, want empty for a top-level chart", i, jobs[i].parent)
+		}
+		if jobs[i].depth != 0 {
+			t.Errorf("jobs[%d].depth = %d, want 0 for a top-level chart", i, jobs[i].depth)
+		}
+	}
+}
+
+// TestResolveChartJobs_ChartOptsNotShared is the regression test for the
+// shallow AddChartOpts copy: every job must own its *action.ChartPathOptions,
+// so mutating one job's RepoURL/Version cannot leak into a sibling's.
+func TestResolveChartJobs_ChartOptsNotShared(t *testing.T) {
+	charts := []v1.Chart{
+		{Name: "rancher", RepoURL: "https://releases.rancher.com/server-charts/stable", Version: "2.9.0"},
+		{Name: "cert-manager", RepoURL: "https://charts.jetstack.io", Version: "1.15.0"},
+		{Name: "longhorn", RepoURL: "https://charts.longhorn.io", Version: "1.7.0"},
+	}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", charts)
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != len(charts) {
+		t.Fatalf("expected %d jobs, got %d", len(charts), len(jobs))
+	}
+
+	for i := range jobs {
+		for k := i + 1; k < len(jobs); k++ {
+			if jobs[i].opts.ChartOpts == jobs[k].opts.ChartOpts {
+				t.Fatalf("jobs[%d] and jobs[%d] share one *action.ChartPathOptions pointee", i, k)
+			}
+		}
+	}
+
+	jobs[0].opts.ChartOpts.RepoURL = "file:///tmp/subchart"
+	jobs[0].opts.ChartOpts.Version = "0.0.0-mutated"
+
+	for i := 1; i < len(jobs); i++ {
+		if got := jobs[i].opts.ChartOpts.RepoURL; got != charts[i].RepoURL {
+			t.Errorf("jobs[%d].opts.ChartOpts.RepoURL = %q, want %q", i, got, charts[i].RepoURL)
+		}
+		if got := jobs[i].opts.ChartOpts.Version; got != charts[i].Version {
+			t.Errorf("jobs[%d].opts.ChartOpts.Version = %q, want %q", i, got, charts[i].Version)
+		}
+	}
+}
+
+func TestResolveChartJobs_NoCharts(t *testing.T) {
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", nil)
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 0 {
+		t.Fatalf("expected no jobs, got %d", len(jobs))
+	}
+}
+
+// TestResolveChartJobs_CredentialFields pins that every TLS/verification
+// field on v1.Chart reaches the job's ChartOpts unchanged.
+func TestResolveChartJobs_CredentialFields(t *testing.T) {
+	ch := v1.Chart{
+		Name:                  "rancher",
+		Verify:                true,
+		Keyring:               "/keys/pubring.gpg",
+		PassCredentialsAll:    true,
+		CertFile:              "/certs/client.crt",
+		KeyFile:               "/certs/client.key",
+		CaFile:                "/certs/ca.crt",
+		InsecureSkipTLSVerify: true,
+		PlainHTTP:             true,
+	}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", []v1.Chart{ch})
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 1 {
+		t.Fatalf("expected 1 job, got %d", len(jobs))
+	}
+
+	opts := jobs[0].opts.ChartOpts
+	if opts.Verify != ch.Verify {
+		t.Errorf("Verify = %v, want %v", opts.Verify, ch.Verify)
+	}
+	if opts.Keyring != ch.Keyring {
+		t.Errorf("Keyring = %q, want %q", opts.Keyring, ch.Keyring)
+	}
+	if opts.PassCredentialsAll != ch.PassCredentialsAll {
+		t.Errorf("PassCredentialsAll = %v, want %v", opts.PassCredentialsAll, ch.PassCredentialsAll)
+	}
+	if opts.CertFile != ch.CertFile {
+		t.Errorf("CertFile = %q, want %q", opts.CertFile, ch.CertFile)
+	}
+	if opts.KeyFile != ch.KeyFile {
+		t.Errorf("KeyFile = %q, want %q", opts.KeyFile, ch.KeyFile)
+	}
+	if opts.CaFile != ch.CaFile {
+		t.Errorf("CaFile = %q, want %q", opts.CaFile, ch.CaFile)
+	}
+	if opts.InsecureSkipTLSVerify != ch.InsecureSkipTLSVerify {
+		t.Errorf("InsecureSkipTLSVerify = %v, want %v", opts.InsecureSkipTLSVerify, ch.InsecureSkipTLSVerify)
+	}
+	if opts.PlainHTTP != ch.PlainHTTP {
+		t.Errorf("PlainHTTP = %v, want %v", opts.PlainHTTP, ch.PlainHTTP)
+	}
+}
+
+// TestResolveChartJobs_CredentialEnv pins that UsernameEnv/PasswordEnv are
+// resolved into ChartOpts.Username/Password via resolveChartCreds.
+func TestResolveChartJobs_CredentialEnv(t *testing.T) {
+	t.Setenv("CHART_USER", "alice")
+	t.Setenv("CHART_PASS", "s3cret")
+
+	ch := v1.Chart{Name: "rancher", UsernameEnv: "CHART_USER", PasswordEnv: "CHART_PASS"}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", []v1.Chart{ch})
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 1 {
+		t.Fatalf("expected 1 job, got %d", len(jobs))
+	}
+	if got := jobs[0].opts.ChartOpts.Username; got != "alice" {
+		t.Errorf("Username = %q, want alice", got)
+	}
+	if got := jobs[0].opts.ChartOpts.Password; got != "s3cret" {
+		t.Errorf("Password = %q, want s3cret", got)
+	}
+}
+
+// TestResolveChartJobs_CredentialEnvMismatch pins the fail-fast contract: a
+// chart with only one of UsernameEnv/PasswordEnv set must abort the whole
+// resolve with no jobs returned, not just skip credentials for that chart.
+func TestResolveChartJobs_CredentialEnvMismatch(t *testing.T) {
+	ch := v1.Chart{Name: "rancher", UsernameEnv: "CHART_USER"}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", []v1.Chart{ch})
+	if err == nil {
+		t.Fatal("expected an error for a chart with usernameEnv set but passwordEnv empty")
+	}
+	if jobs != nil {
+		t.Errorf("expected nil jobs on error, got %d", len(jobs))
+	}
+}
+
+// TestResolveChartJobs_CredentialIsolation is the credential analogue of
+// TestResolveChartJobs_ChartOptsNotShared: each job's Username/Password must
+// come from that chart's own env vars, not a sibling's.
+func TestResolveChartJobs_CredentialIsolation(t *testing.T) {
+	t.Setenv("RANCHER_USER", "rancher-user")
+	t.Setenv("RANCHER_PASS", "rancher-pass")
+	t.Setenv("LONGHORN_USER", "longhorn-user")
+	t.Setenv("LONGHORN_PASS", "longhorn-pass")
+
+	charts := []v1.Chart{
+		{Name: "rancher", UsernameEnv: "RANCHER_USER", PasswordEnv: "RANCHER_PASS"},
+		{Name: "longhorn", UsernameEnv: "LONGHORN_USER", PasswordEnv: "LONGHORN_PASS"},
+	}
+
+	jobs, err := resolveChartJobs(&flags.SyncOpts{}, nil, "/manifests", charts)
+	if err != nil {
+		t.Fatalf("resolveChartJobs: %v", err)
+	}
+	if len(jobs) != 2 {
+		t.Fatalf("expected 2 jobs, got %d", len(jobs))
+	}
+
+	if got := jobs[0].opts.ChartOpts.Username; got != "rancher-user" {
+		t.Errorf("jobs[0].opts.ChartOpts.Username = %q, want rancher-user", got)
+	}
+	if got := jobs[0].opts.ChartOpts.Password; got != "rancher-pass" {
+		t.Errorf("jobs[0].opts.ChartOpts.Password = %q, want rancher-pass", got)
+	}
+	if got := jobs[1].opts.ChartOpts.Username; got != "longhorn-user" {
+		t.Errorf("jobs[1].opts.ChartOpts.Username = %q, want longhorn-user", got)
+	}
+	if got := jobs[1].opts.ChartOpts.Password; got != "longhorn-pass" {
+		t.Errorf("jobs[1].opts.ChartOpts.Password = %q, want longhorn-pass", got)
+	}
+}
+
+// --------------------------------------------------------------------------
+// dedupeImageJobs tests
+// --------------------------------------------------------------------------
+
+// dedupeImageJobs key is (name, platform, excludeExtras); these tests pin
+// which of those differences are real pulls and which are duplicates.
+
+func img(name, platform string, excludeExtras bool) imageJob {
+	return imageJob{
+		img:           v1.Image{Name: name},
+		platform:      platform,
+		excludeExtras: excludeExtras,
+	}
+}
+
+func TestDedupeImageJobs(t *testing.T) {
+	tests := []struct {
+		name string
+		in   []imageJob
+		want []imageJob
+	}{
+		{
+			name: "no jobs",
+			in:   nil,
+			want: nil,
+		},
+		{
+			name: "exact duplicates collapse to the first",
+			in: []imageJob{
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+			},
+			want: []imageJob{img("rancher/rancher:v2.9", "linux/amd64", false)},
+		},
+		{
+			name: "same name, different platform stays separate",
+			in: []imageJob{
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+				img("rancher/rancher:v2.9", "linux/arm64", false),
+			},
+			want: []imageJob{
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+				img("rancher/rancher:v2.9", "linux/arm64", false),
+			},
+		},
+		{
+			name: "same name, different excludeExtras stays separate",
+			in: []imageJob{
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+				img("rancher/rancher:v2.9", "linux/amd64", true),
+			},
+			want: []imageJob{
+				img("rancher/rancher:v2.9", "linux/amd64", false),
+				img("rancher/rancher:v2.9", "linux/amd64", true),
+			},
+		},
+		{
+			name: "different names are never merged",
+			in: []imageJob{
+				img("rancher/rancher:v2.9", "", false),
+				img("rancher/rancher-agent:v2.9", "", false),
+			},
+			want: []imageJob{
+				img("rancher/rancher:v2.9", "", false),
+				img("rancher/rancher-agent:v2.9", "", false),
+			},
+		},
+		{
+			name: "same repository, different tag stays separate",
+			in: []imageJob{
+				img("rancher/rancher:v2.9", "", false),
+				img("rancher/rancher:v2.10", "", false),
+			},
+			want: []imageJob{
+				img("rancher/rancher:v2.9", "", false),
+				img("rancher/rancher:v2.10", "", false),
+			},
+		},
+		{
+			name: "first-seen order survives interleaved duplicates",
+			in: []imageJob{
+				img("c:1", "", false),
+				img("a:1", "", false),
+				img("c:1", "", false),
+				img("b:1", "", false),
+				img("a:1", "", false),
+			},
+			want: []imageJob{
+				img("c:1", "", false),
+				img("a:1", "", false),
+				img("b:1", "", false),
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			got := dedupeImageJobs(tc.in)
+			if len(got) != len(tc.want) {
+				t.Fatalf("got %d jobs, want %d", len(got), len(tc.want))
+			}
+			for i := range tc.want {
+				if got[i].img.Name != tc.want[i].img.Name ||
+					got[i].platform != tc.want[i].platform ||
+					got[i].excludeExtras != tc.want[i].excludeExtras {
+					t.Errorf("job[%d] = (%q, %q, %v), want (%q, %q, %v)",
+						i, got[i].img.Name, got[i].platform, got[i].excludeExtras,
+						tc.want[i].img.Name, tc.want[i].platform, tc.want[i].excludeExtras)
+				}
+			}
+		})
+	}
+}
+
+// TestDedupeImageJobs_KeepsFirstOccurrenceFields pins "first occurrence wins"
+// on the whole job, not just its key: a later duplicate's other fields (here,
+// rewrite) must not overwrite the retained job's.
+func TestDedupeImageJobs_KeepsFirstOccurrenceFields(t *testing.T) {
+	first := img("rancher/rancher:v2.9", "linux/amd64", false)
+	first.rewrite = "mirror/rancher"
+
+	second := img("rancher/rancher:v2.9", "linux/amd64", false)
+	second.rewrite = "other/rancher"
+
+	got := dedupeImageJobs([]imageJob{first, second})
+	if len(got) != 1 {
+		t.Fatalf("expected 1 job, got %d", len(got))
+	}
+	if got[0].rewrite != "mirror/rancher" {
+		t.Errorf("rewrite = %q, want mirror/rancher", got[0].rewrite)
+	}
+}
+
+// TestDedupeImageJobs_DoesNotMutateInput pins that callers keep a usable
+// slice: deduping returns a new one rather than compacting in place.
+func TestDedupeImageJobs_DoesNotMutateInput(t *testing.T) {
+	in := []imageJob{
+		img("a:1", "", false),
+		img("b:1", "", false),
+		img("a:1", "", false),
+	}
+
+	_ = dedupeImageJobs(in)
+
+	if len(in) != 3 {
+		t.Fatalf("input length changed to %d", len(in))
+	}
+	for i, want := range []string{"a:1", "b:1", "a:1"} {
+		if in[i].img.Name != want {
+			t.Errorf("in[%d] = %q, want %q", i, in[i].img.Name, want)
+		}
+	}
+}
+
+// --------------------------------------------------------------------------
+// traverseChartLevels graph tests
+//
+// These cover the traversal in isolation: the seen-set cycle guard, the
+// maxChartDepth cap, per-level dependency deduplication, and error
+// propagation. Every test injects a chartFetcher, so none of them touch helm,
+// the store, or the network.
+// --------------------------------------------------------------------------
+
+// newChartJob builds a top-level chartJob for a bare chart name, with the
+// per-job *action.ChartPathOptions allocation resolveChartJobs guarantees.
+func newChartJob(name string) chartJob {
+	return chartJob{
+		cfg:  v1.Chart{Name: name},
+		opts: flags.AddChartOpts{ChartOpts: &action.ChartPathOptions{}},
+	}
+}
+
+// recordingFetcher is a chartFetcher that records the name of every chart it
+// is asked to fetch and returns the dependencies graph[name] declares.
+type recordingFetcher struct {
+	mu    sync.Mutex
+	seen  []string
+	graph map[string][]string
+}
+
+func (r *recordingFetcher) fetch(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+	r.mu.Lock()
+	r.seen = append(r.seen, j.cfg.Name)
+	deps := r.graph[j.cfg.Name]
+	r.mu.Unlock()
+
+	out := make([]chartJob, 0, len(deps))
+	for _, d := range deps {
+		out = append(out, newChartJob(d))
+	}
+	return nil, out, nil
+}
+
+func (r *recordingFetcher) count(name string) int {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	n := 0
+	for _, s := range r.seen {
+		if s == name {
+			n++
+		}
+	}
+	return n
+}
+
+// TestTraverseChartLevels_TerminatesOnCycle pins the seen-set cycle guard,
+// without which the walk recurses forever: A depends on B, B depends back on
+// A, and each must be fetched exactly once.
+func TestTraverseChartLevels_TerminatesOnCycle(t *testing.T) {
+	f := &recordingFetcher{graph: map[string][]string{
+		"a": {"b"},
+		"b": {"a"},
+	}}
+
+	_, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("a")}, 2, f.fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+
+	if got := f.count("a"); got != 1 {
+		t.Errorf("chart a fetched %d times, want exactly 1", got)
+	}
+	if got := f.count("b"); got != 1 {
+		t.Errorf("chart b fetched %d times, want exactly 1", got)
+	}
+	if charts != 2 {
+		t.Errorf("charts fetched = %d, want 2", charts)
+	}
+}
+
+// TestTraverseChartLevels_SelfCycleTerminates covers the degenerate cycle a
+// chart declaring itself as a dependency would produce.
+func TestTraverseChartLevels_SelfCycleTerminates(t *testing.T) {
+	f := &recordingFetcher{graph: map[string][]string{"a": {"a"}}}
+
+	_, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("a")}, 2, f.fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+	if charts != 1 {
+		t.Errorf("charts fetched = %d, want 1", charts)
+	}
+}
+
+// TestTraverseChartLevels_SeenKeyIncludesRepoAndVersion asserts the seen set
+// is keyed on name|repoURL|version rather than name alone: the same chart
+// name at two versions is two genuinely different pulls.
+func TestTraverseChartLevels_SeenKeyIncludesRepoAndVersion(t *testing.T) {
+	var mu sync.Mutex
+	var fetched []string
+
+	fetch := func(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+		mu.Lock()
+		fetched = append(fetched, j.cfg.Name+"@"+j.cfg.Version)
+		mu.Unlock()
+
+		if j.cfg.Version != "" {
+			return nil, nil, nil
+		}
+		return nil, []chartJob{
+			{cfg: v1.Chart{Name: "dep", RepoURL: "https://charts.example.com", Version: "1.0.0"}, opts: flags.AddChartOpts{ChartOpts: &action.ChartPathOptions{}}},
+			{cfg: v1.Chart{Name: "dep", RepoURL: "https://charts.example.com", Version: "2.0.0"}, opts: flags.AddChartOpts{ChartOpts: &action.ChartPathOptions{}}},
+		}, nil
+	}
+
+	_, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("parent")}, 2, fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+	if charts != 3 {
+		t.Fatalf("charts fetched = %d, want 3 (parent + dep@1.0.0 + dep@2.0.0), got %v", charts, fetched)
+	}
+}
+
+// TestTraverseChartLevels_DedupesRepeatedDependenciesInOneLevel covers two
+// siblings declaring the identical dependency: it is fetched once, not twice.
+func TestTraverseChartLevels_DedupesRepeatedDependenciesInOneLevel(t *testing.T) {
+	f := &recordingFetcher{graph: map[string][]string{
+		"a":      {"shared"},
+		"b":      {"shared"},
+		"shared": nil,
+	}}
+
+	_, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("a"), newChartJob("b")}, 2, f.fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+	if got := f.count("shared"); got != 1 {
+		t.Errorf("shared dependency fetched %d times, want exactly 1", got)
+	}
+	if charts != 3 {
+		t.Errorf("charts fetched = %d, want 3", charts)
+	}
+}
+
+// TestTraverseChartLevels_DepthCapHolds pins maxChartDepth independently of
+// the seen set: an infinitely deep chain of distinct charts (which the seen
+// set can never stop) must terminate at exactly maxChartDepth levels.
+func TestTraverseChartLevels_DepthCapHolds(t *testing.T) {
+	var mu sync.Mutex
+	n := 0
+
+	fetch := func(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+		mu.Lock()
+		n++
+		mu.Unlock()
+		return nil, []chartJob{newChartJob(j.cfg.Name + "-child")}, nil
+	}
+
+	_, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("root")}, 2, fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+
+	mu.Lock()
+	defer mu.Unlock()
+	if n != maxChartDepth {
+		t.Errorf("fetch called %d times, want exactly maxChartDepth (%d) -- one chart per level on an unbounded chain", n, maxChartDepth)
+	}
+	if charts != maxChartDepth {
+		t.Errorf("charts fetched = %d, want %d", charts, maxChartDepth)
+	}
+}
+
+// TestTraverseChartLevels_CollectsImagesAcrossLevels asserts images
+// discovered at every depth reach the caller, not just the top level's.
+func TestTraverseChartLevels_CollectsImagesAcrossLevels(t *testing.T) {
+	fetch := func(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+		imgs := []imageJob{{img: v1.Image{Name: j.cfg.Name + "-image"}}}
+		if j.cfg.Name == "parent" {
+			return imgs, []chartJob{newChartJob("child")}, nil
+		}
+		return imgs, nil, nil
+	}
+
+	images, charts, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("parent")}, 2, fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+	if charts != 2 {
+		t.Fatalf("charts fetched = %d, want 2", charts)
+	}
+
+	got := map[string]bool{}
+	for _, i := range images {
+		got[i.img.Name] = true
+	}
+	for _, want := range []string{"parent-image", "child-image"} {
+		if !got[want] {
+			t.Errorf("image %q missing from discovered set %v", want, got)
+		}
+	}
+}
+
+// TestTraverseChartLevels_PropagatesFetchError asserts a fetch failure
+// surfaces verbatim rather than as an errgroup cancellation, matching
+// runImageJobs' semantics.
+func TestTraverseChartLevels_PropagatesFetchError(t *testing.T) {
+	sentinel := errors.New("chart repo unreachable")
+
+	fetch := func(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+		if j.cfg.Name == "bad" {
+			return nil, nil, sentinel
+		}
+		return nil, nil, nil
+	}
+
+	_, _, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("bad")}, 1, fetch)
+	if !errors.Is(err, sentinel) {
+		t.Fatalf("traverseChartLevels error = %v, want it to wrap %v", err, sentinel)
+	}
+}
+
+// TestTraverseChartLevels_ErrorInDeeperLevelStopsTraversal asserts a failure
+// discovered at depth 1 aborts before any depth-2 chart is fetched.
+func TestTraverseChartLevels_ErrorInDeeperLevelStopsTraversal(t *testing.T) {
+	sentinel := errors.New("subchart missing")
+	f := &recordingFetcher{graph: map[string][]string{
+		"root": {"mid"},
+		"mid":  {"leaf"},
+	}}
+
+	fetch := func(ctx context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+		if j.cfg.Name == "mid" {
+			return nil, nil, sentinel
+		}
+		return f.fetch(ctx, j)
+	}
+
+	if _, _, err := traverseChartLevels(context.Background(), []chartJob{newChartJob("root")}, 1, fetch); !errors.Is(err, sentinel) {
+		t.Fatalf("traverseChartLevels error = %v, want it to wrap %v", err, sentinel)
+	}
+	if got := f.count("leaf"); got != 0 {
+		t.Errorf("leaf fetched %d times after its parent level failed, want 0", got)
+	}
+}
+
+// TestTraverseChartLevels_NoJobs covers the empty-input path.
+func TestTraverseChartLevels_NoJobs(t *testing.T) {
+	fetch := func(context.Context, chartJob) ([]imageJob, []chartJob, error) {
+		return nil, nil, fmt.Errorf("fetch must not be called with no jobs")
+	}
+
+	images, charts, err := traverseChartLevels(context.Background(), nil, 4, fetch)
+	if err != nil {
+		t.Fatalf("traverseChartLevels: %v", err)
+	}
+	if len(images) != 0 || charts != 0 {
+		t.Errorf("images = %v, charts = %d, want none", images, charts)
+	}
+}
+
+// --------------------------------------------------------------------------
+// Concurrent chart pipeline tests
+//
+// These cover that traverseChartLevels honors its concurrency bound, that
+// dependency-derived jobs never share an *action.ChartPathOptions with their
+// parent, and that a file:// dependency still resolves after its parent's
+// level has completed -- the regression that runChartJobs' single shared temp
+// root exists to prevent.
+// --------------------------------------------------------------------------
+
+// newFileDependencyChartJob builds a top-level chartJob for
+// testdata/chart-with-file-dependency-chart-1.0.0.tgz, whose Chart.yaml
+// declares two dependencies that resolve inside the parent's own expanded
+// directory (one via file://, one via an empty repository field).
+func newFileDependencyChartJob() chartJob {
+	return chartJob{
+		cfg: v1.Chart{Name: "chart-with-file-dependency-chart-1.0.0.tgz", RepoURL: chartTestdataDir},
+		opts: flags.AddChartOpts{
+			ChartOpts:       &action.ChartPathOptions{RepoURL: chartTestdataDir},
+			AddDependencies: true,
+		},
+	}
+}
+
+// TestTraverseChartLevels_BoundedFanOut asserts the observed peak of
+// simultaneously in-flight fetches never exceeds the requested concurrency,
+// at every level of the walk.
+func TestTraverseChartLevels_BoundedFanOut(t *testing.T) {
+	const perLevel = 12
+
+	for _, concurrency := range []int{1, 2, 4} {
+		t.Run(fmt.Sprintf("concurrency=%d", concurrency), func(t *testing.T) {
+			var (
+				mu       sync.Mutex
+				inFlight int
+				peak     int
+			)
+
+			fetch := func(_ context.Context, j chartJob) ([]imageJob, []chartJob, error) {
+				mu.Lock()
+				inFlight++
+				if inFlight > peak {
+					peak = inFlight
+				}
+				mu.Unlock()
+
+				// Long enough that a broken bound would overlap observably;
+				// short enough to keep the whole table under a second.
+				time.Sleep(2 * time.Millisecond)
+
+				mu.Lock()
+				inFlight--
+				mu.Unlock()
+
+				if strings.HasSuffix(j.cfg.Name, "-child") {
+					return nil, nil, nil
+				}
+				return nil, []chartJob{newChartJob(j.cfg.Name + "-child")}, nil
+			}
+
+			jobs := make([]chartJob, 0, perLevel)
+			for i := 0; i < perLevel; i++ {
+				jobs = append(jobs, newChartJob(fmt.Sprintf("chart%d", i)))
+			}
+
+			_, charts, err := traverseChartLevels(context.Background(), jobs, concurrency, fetch)
+			if err != nil {
+				t.Fatalf("traverseChartLevels: %v", err)
+			}
+			if charts != 2*perLevel {
+				t.Errorf("charts fetched = %d, want %d", charts, 2*perLevel)
+			}
+
+			mu.Lock()
+			defer mu.Unlock()
+			if peak > concurrency {
+				t.Errorf("peak in-flight fetches = %d, want <= %d", peak, concurrency)
+			}
+			if peak < 1 {
+				t.Errorf("peak in-flight fetches = %d, want at least 1", peak)
+			}
+		})
+	}
+}
+
+// TestFetchChart_DependencyJobsDoNotShareChartOpts is the pointer-identity
+// regression test for the derived-job path. resolveChartJobs' equivalent test
+// covers only top-level jobs; a `depJob := parentJob` struct copy would pass
+// that one while still sharing the *action.ChartPathOptions pointee here,
+// which under concurrency is a live data race on RepoURL/Version.
+func TestFetchChart_DependencyJobsDoNotShareChartOpts(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+	rso := defaultRootOpts(s.Root)
+	ro := defaultCliOpts()
+
+	parent := newFileDependencyChartJob()
+	_, deps, err := fetchChart(ctx, s, parent, t.TempDir(), rso, ro)
+	if err != nil {
+		t.Fatalf("fetchChart: %v", err)
+	}
+	if len(deps) != 2 {
+		t.Fatalf("dependency jobs = %d, want 2 (child + crds)", len(deps))
+	}
+
+	for i, d := range deps {
+		if d.opts.ChartOpts == parent.opts.ChartOpts {
+			t.Fatalf("deps[%d] shares its parent's *action.ChartPathOptions pointee", i)
+		}
+	}
+	if deps[0].opts.ChartOpts == deps[1].opts.ChartOpts {
+		t.Fatal("deps[0] and deps[1] share one *action.ChartPathOptions pointee")
+	}
+
+	parentRepoURL := parent.opts.ChartOpts.RepoURL
+	deps[0].opts.ChartOpts.RepoURL = "https://mutated.example.com"
+	deps[0].opts.ChartOpts.Version = "0.0.0-mutated"
+
+	if got := parent.opts.ChartOpts.RepoURL; got != parentRepoURL {
+		t.Errorf("parent RepoURL = %q after mutating a dependency's, want %q", got, parentRepoURL)
+	}
+	if got := deps[1].opts.ChartOpts.RepoURL; got == "https://mutated.example.com" {
+		t.Error("mutating deps[0] was visible through deps[1]")
+	}
+
+	parent.opts.ChartOpts.Version = "9.9.9-mutated"
+	for i, d := range deps {
+		if d.opts.ChartOpts.Version == "9.9.9-mutated" {
+			t.Errorf("mutating the parent was visible through deps[%d]", i)
+		}
+	}
+}
+
+// TestFetchChart_DerivedDependencyFields pins the shape of a derived job:
+// dependencies are walked further, images are not rediscovered per subchart,
+// no rewrite is inherited, and parent/depth are set for attribution and the
+// depth cap.
+func TestFetchChart_DerivedDependencyFields(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+	rso := defaultRootOpts(s.Root)
+	ro := defaultCliOpts()
+
+	parent := newFileDependencyChartJob()
+	parent.opts.AddImages = true
+	parent.rewrite = "myorg/parent-chart"
+
+	_, deps, err := fetchChart(ctx, s, parent, t.TempDir(), rso, ro)
+	if err != nil {
+		t.Fatalf("fetchChart: %v", err)
+	}
+	if len(deps) != 2 {
+		t.Fatalf("dependency jobs = %d, want 2", len(deps))
+	}
+
+	for i, d := range deps {
+		if !d.opts.AddDependencies {
+			t.Errorf("deps[%d].opts.AddDependencies = false, want true", i)
+		}
+		if d.opts.AddImages {
+			t.Errorf("deps[%d].opts.AddImages = true, want false: the parent's render already covered the whole tree", i)
+		}
+		if d.rewrite != "" {
+			t.Errorf("deps[%d].rewrite = %q, want empty", i, d.rewrite)
+		}
+		if d.depth != parent.depth+1 {
+			t.Errorf("deps[%d].depth = %d, want %d", i, d.depth, parent.depth+1)
+		}
+		if !strings.Contains(d.parent, "chart-with-file-dependency-chart") {
+			t.Errorf("deps[%d].parent = %q, want it to name the parent chart's ref", i, d.parent)
+		}
+	}
+}
+
+// TestRunChartJobs_FileDependencyResolvesAfterParentLevel is the
+// shared-temp-root regression test. A file:// dependency's chartJob names a
+// path *inside* its parent's expanded directory, and BFS fetches it only
+// after the parent's whole level has returned. A per-chart temp dir removed
+// when fetchChart returns would delete that path out from under the child, so
+// this test fails outright under that design rather than flaking.
+func TestRunChartJobs_FileDependencyResolvesAfterParentLevel(t *testing.T) {
+	for _, concurrency := range []int{1, 4} {
+		t.Run(fmt.Sprintf("concurrency=%d", concurrency), func(t *testing.T) {
+			ctx := newTestContext(t)
+			s := newTestStore(t)
+			rso := defaultRootOpts(s.Root)
+			ro := defaultCliOpts()
+
+			if err := runChartJobs(ctx, s, []chartJob{newFileDependencyChartJob()}, concurrency, rso, ro, nil); err != nil {
+				t.Fatalf("runChartJobs: %v", err)
+			}
+
+			assertArtifactInStore(t, s, "chart-with-file-dependency-chart:1.0.0")
+			assertArtifactInStore(t, s, "child:2.0.0")
+			assertArtifactInStore(t, s, "crds:0.0.1")
+		})
+	}
+}
+
+// TestRunChartJobs_RemovesTempRoot asserts the shared temp root does not
+// outlive the call: it holds every chart's expansion, so leaking one per
+// `store sync` invocation would accumulate on disk.
+func TestRunChartJobs_RemovesTempRoot(t *testing.T) {
+	ctx := newTestContext(t)
+	s := newTestStore(t)
+	rso := defaultRootOpts(s.Root)
+	rso.TempOverride = t.TempDir()
+	ro := defaultCliOpts()
+
+	if err := runChartJobs(ctx, s, []chartJob{newFileDependencyChartJob()}, 2, rso, ro, nil); err != nil {
+		t.Fatalf("runChartJobs: %v", err)
+	}
+
+	entries, err := os.ReadDir(rso.TempOverride)
+	if err != nil {
+		t.Fatalf("ReadDir %s: %v", rso.TempOverride, err)
+	}
+	for _, e := range entries {
+		t.Errorf("temp override still contains %q after runChartJobs returned", filepath.Join(rso.TempOverride, e.Name()))
+	}
+}
+
+// TestFormatAddedLine_NilStats proves formatAddedLine never prints
+// "0 layers" when stats is nil, and produces a sane elapsed-only line.
+func TestFormatAddedLine_NilStats(t *testing.T) {
+	got := formatAddedLine("example.com/repo:v1", nil, 1500*time.Millisecond)
+
+	if strings.Contains(got, "0 layer") {
+		t.Errorf("formatAddedLine with nil stats must never print \"0 layers\", got %q", got)
+	}
+	if !strings.Contains(got, "example.com/repo:v1") {
+		t.Errorf("formatAddedLine must include the ref, got %q", got)
+	}
+	if !strings.Contains(got, "1.5s") {
+		t.Errorf("formatAddedLine must include the elapsed time as %%.1fs, got %q", got)
+	}
+	if !strings.HasPrefix(got, "✓ added") {
+		t.Errorf("formatAddedLine must start with \"✓ added\", got %q", got)
+	}
+}
+
+// TestFormatAddedLine_ZeroValueStats proves formatAddedLine treats a
+// zero-value *store.ImageStats (Layers == 0) the same as nil: elapsed-only,
+// never "0 layers".
+func TestFormatAddedLine_ZeroValueStats(t *testing.T) {
+	stats := &store.ImageStats{}
+	got := formatAddedLine("example.com/repo:v1", stats, 2*time.Second)
+
+	if strings.Contains(got, "0 layer") {
+		t.Errorf("formatAddedLine with zero-value stats must never print \"0 layers\", got %q", got)
+	}
+	if !strings.Contains(got, "2.0s") {
+		t.Errorf("formatAddedLine must include the elapsed time as %%.1fs, got %q", got)
+	}
+}
+
+// TestFormatAddedLine_WithStats proves formatAddedLine includes layer
+// count (correctly pluralized), human-readable byte size, and elapsed time
+// when stats has at least one layer.
+func TestFormatAddedLine_WithStats(t *testing.T) {
+	tests := []struct {
+		name    string
+		layers  int64
+		bytes   int64
+		wantSub string
+	}{
+		{name: "singular layer", layers: 1, bytes: 100, wantSub: "1 layer,"},
+		{name: "plural layers", layers: 3, bytes: 1_500_000, wantSub: "3 layers,"},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			stats := &store.ImageStats{}
+			stats.Layers.Store(tt.layers)
+			stats.Bytes.Store(tt.bytes)
+
+			got := formatAddedLine("example.com/repo:v1", stats, 3*time.Second)
+
+			if !strings.Contains(got, tt.wantSub) {
+				t.Errorf("formatAddedLine = %q, want substring %q", got, tt.wantSub)
+			}
+			wantSize := humanize.Bytes(uint64(tt.bytes))
+			if !strings.Contains(got, wantSize) {
+				t.Errorf("formatAddedLine = %q, want it to contain human size %q", got, wantSize)
+			}
+			if !strings.Contains(got, "3.0s") {
+				t.Errorf("formatAddedLine = %q, want it to contain elapsed \"3.0s\"", got)
+			}
+		})
+	}
 }
